refactor(wash-cli): refactor the code for wash dev
