+++ crates/wash-cli/src/bin/wash.rs
@@ -26,12 +26,12 @@ use wash_cli::app::{self, AppCliCommand};
 use wash_cli::build::{self, BuildCommand};
 use wash_cli::call::{self, CallCli};
 use wash_cli::cmd::config::{self, ConfigCliCommand};
+use wash_cli::cmd::dev::{self, DevCommand};
 use wash_cli::cmd::up::{self, UpCommand};
 use wash_cli::common;
 use wash_cli::completions::{self, CompletionOpts};
 use wash_cli::config::{NATS_SERVER_VERSION, WADM_VERSION, WASMCLOUD_HOST_VERSION};
 use wash_cli::ctx::{self, CtxCommand};
-use wash_cli::dev::{self, DevCommand};
 use wash_cli::down::{self, DownCommand};
 use wash_cli::drain;
 use wash_cli::generate::{self, NewCliCommand};
