+++ crates/wash-cli/src/cmd/mod.rs
@@ -1,4 +1,5 @@
 //! Commands that are exposed by `wash`
 
 pub mod config;
+pub mod dev;
 pub mod up;
